Create set-roles.js; Modify authenticateUser, createClothe, getUserUsername, registerUser and removeClothe #98

+++ staff/Laura-Anton/project/api/logic/removeClothe.js
@@ -1,12 +1,7 @@
 import { User, Clothe } from '../data/index.js'
-import { validate, SystemError, NotFoundError, AuthorshipError } from 'com'
+import { validate, SystemError, NotFoundError, AuthorshipError, RoleError } from 'com'
 
-/**
- * Removes a post by id from database.
- * 
- * @param {string} userId The requester user id.
- * @param {string} clotheId The clothe id to remove.
- */
+
 export const removeClothe = (userId, clotheId) => {
     validate.id(userId)
     validate.id(clotheId)
@@ -15,14 +10,14 @@ export const removeClothe = (userId, clotheId) => {
         .catch(error => { throw new SystemError('mongo error') })
         .then(user => {
             if (!user) throw new NotFoundError('user not found')
+            if (user.role !== 'administrator') throw new RoleError('user is not administrator') 
+            
 
             return Clothe.findById(clotheId)
                 .catch(error => { throw new SystemError('mongo error') })
                 .then(clothe => {
                     if (!clothe) throw new NotFoundError('clothe not found')
 
-                    //if (user.role !== 'administrator') throw new AuthorshipError('user not author of clothe')
-//TODO asignar administrador cuando esten los videos
                     return Clothe.deleteOne({ _id: clotheId })
                         .catch(error => { throw new SystemError('mongo error') })
                         .then(() => { })
